+++ lib/cassandra_migrations/migration/table_operations.rb
@@ -21,6 +21,7 @@ module CassandraMigrations
         table_definition = TableDefinition.new
         table_definition.define_primary_keys(options[:primary_keys]) if options[:primary_keys]
         table_definition.define_partition_keys(options[:partition_keys]) if options[:partition_keys]
+        table_definition.define_options(options[:options]) if options[:options]
 
         yield table_definition if block_given?
 
@@ -29,21 +30,7 @@ module CassandraMigrations
         create_cql =  "CREATE TABLE #{table_name} ("
         create_cql << table_definition.to_create_cql
         create_cql << ")"
-
-        cql_options_clause = []
-
-        if options[:clustering_order]
-          cql_options_clause << "CLUSTERING ORDER BY (#{options[:clustering_order]})"
-        end
-
-        if options[:options]
-          options[:options].each {|name, value|
-            value_str = value.map {|k, v| "'#{k}':'#{v}'"}.join(',')
-            cql_options_clause << "#{name} = {#{value_str}}"
-          }
-        end
-
-        create_cql << " WITH #{cql_options_clause.join ' AND '}" if !cql_options_clause.empty?
+        create_cql << table_definition.options
 
         announce_suboperation create_cql