+++ spec/cassandra_migrations/config_spec.rb
@@ -4,20 +4,20 @@ require 'spec_helper'
 describe CassandraMigrations::Config do
 
   before do
-    CassandraMigrations::Config.config = nil
+    CassandraMigrations::Config.configurations = nil
   end
 
   it 'should fetch values in config for the right environment' do
-    Rails.stub(:root).and_return Pathname.new("spec/fixtures")
-    Rails.stub(:env).and_return ActiveSupport::StringInquirer.new("development")
+    allow(Rails).to receive(:root).and_return Pathname.new("spec/fixtures")
+    allow(Rails).to receive(:env).and_return ActiveSupport::StringInquirer.new("development")
 
-    CassandraMigrations::Config.keyspace.should == 'cassandra_migrations_development'
-    CassandraMigrations::Config.replication.should == {'class' => "SimpleStrategy", 'replication_factor' => 1 }
+    expect(CassandraMigrations::Config.keyspace).to eq('cassandra_migrations_development')
+    expect(CassandraMigrations::Config.replication).to eq({'class' => "SimpleStrategy", 'replication_factor' => 1 })
   end
 
   it 'should raise exception if there are no configs for environment' do
-    Rails.stub(:root).and_return Pathname.new("spec/fixtures")
-    Rails.stub(:env).and_return ActiveSupport::StringInquirer.new("fake_environment")
+    allow(Rails).to receive(:root).and_return Pathname.new("spec/fixtures")
+    allow(Rails).to receive(:env).and_return ActiveSupport::StringInquirer.new("fake_environment")
 
     expect {
       CassandraMigrations::Config.keyspace
@@ -25,8 +25,8 @@ describe CassandraMigrations::Config do
   end
 
   it 'should raise exception if there are no config file' do
-    Rails.stub(:root).and_return Pathname.new("spec/fake_path")
-    Rails.stub(:env).and_return ActiveSupport::StringInquirer.new("development")
+    allow(Rails).to receive(:root).and_return Pathname.new("spec/fake_path")
+    allow(Rails).to receive(:env).and_return ActiveSupport::StringInquirer.new("development")
 
     expect {
       CassandraMigrations::Config.keyspace
@@ -34,14 +34,20 @@ describe CassandraMigrations::Config do
   end
 
   it 'should allow ERB in the config file' do
-    Rails.stub(:root).and_return Pathname.new("spec/fixtures/with_erb")
-    Rails.stub(:env).and_return ActiveSupport::StringInquirer.new("test")
+    allow(Rails).to receive(:root).and_return Pathname.new("spec/fixtures/with_erb")
+    allow(Rails).to receive(:env).and_return ActiveSupport::StringInquirer.new("test")
 
-    CassandraMigrations::Config.keyspace.should == 'cassandra_migrations_test'
+    expect(CassandraMigrations::Config.keyspace).to eq('cassandra_migrations_test')
 
-    CassandraMigrations::Config.config = nil
-    ENV.stub(:[]).with("CI").and_return("true")
+    CassandraMigrations::Config.configurations = nil
+    allow(ENV).to receive(:[]).with("CI").and_return("true")
 
-    CassandraMigrations::Config.keyspace.should == 'cassandra_migrations_ci'
+    expect(CassandraMigrations::Config.keyspace).to eq('cassandra_migrations_ci')
+  end
+
+  it 'allows access to configurations for other environments than the current Rails.env' do
+    allow(Rails).to receive(:root).and_return Pathname.new("spec/fixtures")
+    allow(Rails).to receive(:env).and_return ActiveSupport::StringInquirer.new("development")
+    expect(CassandraMigrations::Config.configurations['production'].keyspace).to eq('cassandra_migrations_production')
   end
 end